Add a version which fixes the wall bug
This version uses the environment and the markDetectedWalls function to update the detectedMap properly because
the walls that the drone was passing through had the value 0 in the matrix. Therefore I assume this might be some order-related bug regarding the way the position of
the drone is updated.
The backtracking was intentionally left out because this bug was pretty serious.
TODO: Find the exact reason why the program didn't work before and remove the reference to the Environment (e) parameter