This is a combination of 7 commits.
Fix crossover function for two individuals

The complementary probability for the crossover is also taken into account, so there is a small chance(0.2) that the two individuals will not be able to reproduce.

Fix fitness function by removing unintended position reset

The currentPosition in the for loop of the fitness function used to be reset back to the starting position, which was obviously incorrect.

Fix chromosome gene/integer bug in the mutation function

The mutation function used to replace a gene from the chromosome with a 0/1/2/3 which wasn't wrapped inside a Gene instance.

Ensure that crossover and mutation are correct

Fix visualised surface-marking function

The current position of the drone was not used. Dumb mistake.

Sort individuals by their fitness before selection

I can't believe I forgot this detail...

Implement pure reinsertion

The parents are replaced by their offspring. In addition, the less-fit individuals will be allowed to reproduce, instead of having a fitness score of 0 assigned to them.